Fixed GNU warning.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@34142 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/ptlib/common/vconvert.cxx
@@ -1885,7 +1885,7 @@ bool PStandardColourConverter::YUV420PtoRGB(const BYTE * srcFrameBuffer,
 #endif
 
   BYTE * scanLinePtrRGB = dstFrameBuffer;
-  int scanLineSizeRGB = (rgbIncrement*m_dstFrameWidth+3)&~3;
+  int scanLineSizeRGB = (int)((rgbIncrement*m_dstFrameWidth+3)&~3);
 
   unsigned srcPixpos[4] = { 0, 1, m_srcFrameWidth, m_srcFrameWidth + 1 };
   unsigned dstPixpos[4] = { 0, rgbIncrement, scanLineSizeRGB, scanLineSizeRGB+rgbIncrement };
